+++ models/CreateDocument.php
@@ -66,7 +66,7 @@ class CreateDocument extends Model
             $source = $this->templatePath() . '/new.' . $this->extension;
             $newFile = $this->fileName . '.' . $this->extension;
 
-            $mime = $module->mimes[$this->extension];
+            $mime = $module->formats()->mimes[$this->extension];
 
             $file = new File();
             $file->file_name = $newFile;
@@ -87,7 +87,7 @@ class CreateDocument extends Model
 
         $lang = ($user) && !empty($user->language) ? $user->language : Yii::$app->language;
         if (!array_key_exists($lang, $module->languageCodes)) {
-            $lang = 'en-US';
+            $lang = 'default';
         }
 
         return $module->getAssetPath() . '/templates/' . $module->languageCodes[$lang];