+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
+## Added
+- default empty templates
+- ar-SA, eu-ES, fi-FI, he-IL, nb-NO, sl-SL and sr-Cyrl-RS empty file templates
+
 ## Changed
+- creating and editing pdf form
+- updatable list of supported formats
 - demo server address changed
 
 ## 3.1.0